backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-nmap==0.7.1
python-multipart==0.0.6
pydantic==2.5.0 